Move array_ufunc to directories
